requirements.txt
IProgress==0.4
ipykernel==7.2.0
kaleido==1.2.0
markdown-it-py==4.0.0
matplotlib==3.10.8
nbformat==5.10.4
optuna==4.8.0
pandas==3.0.1
pip==26.0.1
plotly==6.6.0
scikit-image==0.26.0
scikit-learn==1.8.0
torchvision==0.25.0+cu130
wandb==0.25.1
wheel==0.46.3
